refactored for exporting game as an object
